Separate calculateInterval core logic into shared pkg
