+++ packages/functions/src/utils.ts
@@ -17,6 +17,7 @@ import {
   OrgSubCollection,
   SlotType,
   SlotsByDay,
+  calculateIntervalDuration,
 } from "@eisbuk/shared";
 
 type Auth = CallableContext["auth"];
@@ -340,42 +341,6 @@ export const validateJSON = <T extends Record<string, any>>(
   return [null, constructValidationErrors(validate.errors || [], prefix), {}];
 };
 // #endregion JSONValidation
-
-// #region CalculateInterval
-
-/**
- * @param {string | null} interval - String slot interval
- * Converts a string slot interval to a number e.g:
- * `null => 0`;
- * `"21:00 - 21:20" => 0.5`;
- * `"16:00 - 17:00" => 1.0`;
- * `"22:00 - 23:30" => 1.5`;
- * `"22:00 - 24:00" => 2`;
- * `"22:20 - 24:00" => 2`;
- */
-export const calculateIntervalDuration = (interval: string | null) => {
-  if (interval === null) {
-    return 0;
-  }
-
-  const [startTime, endTime] = interval.split("-");
-  const hourInMillis = 3600000;
-
-  const diffMillis =
-    getMillisFromMidnight(endTime) - getMillisFromMidnight(startTime);
-
-  return Math.ceil((diffMillis / hourInMillis) * 2) * 0.5;
-};
-
-/**
- * Calculate milliseconds passed from start of day (for ISO time string, eg. "09:00")
- */
-const getMillisFromMidnight = (time: string) =>
-  time
-    .split(":")
-    .reduce((acc, curr, i) => acc + parseInt(curr) * 1000 * 60 ** (2 - i), 0);
-
-// #endregion CalculateInterval
 
 /**
  * Calculates the total duration of booked slots for a customer, categorized by month and slot type.
