+++ packages/ui/src/AttendanceSheet/AttendanceSheet.tsx
@@ -8,7 +8,7 @@ import i18n, {
 } from "@eisbuk/translations";
 import { CustomerWithAttendance, SlotType } from "@eisbuk/shared";
 
-import { calculateIntervalDuration } from "./utils";
+import { getIntervalString } from "./utils";
 import Table from "../Table";
 
 interface RowItem {
@@ -130,7 +130,7 @@ const processTableData = (entries: TableDataEntry[]): RowItem[] =>
       .filter(({ bookedInterval }) => Boolean(bookedInterval))
       .map(({ name, surname, bookedInterval }) => {
         const [start, end] = bookedInterval!.split("-");
-        const totalHours = calculateIntervalDuration(start, end);
+        const totalHours = getIntervalString(bookedInterval!);
 
         return {
           type,
